refactor (cli): move cdsctl workflow command to other files
